Big cleanup, got rid of base
The previous design was flawed.  This design is much better.  It requires an absolute
URL from the entrypoint and then maintains an absolute URL in every nav state.  The
only place where it is even possible to introduce a relative URL is when the SirenNav
instance gets created or when a goto call is made.  The SirenNav creation now throws
and error if you don't include an absolute URL and goto leverages the current
NavStates absolute URL to formulate new absolute URLs.  All other URLs (follow, action,
etc.) are handled internally and also ensure the new NavState has an absolute URL.

Because the URL is always absolute, we don't even need to worry about providing a
BaseURL for the underlying axios config.