Changed the semantics of Accept header handling
Previously, calling `accept` would *add* a new acceptable
content type.  But there was no way to remove, reorder, *etc*.
I thought it was probably better to simply give the user complete
and direct control of the value (as it is with all other headers).
In this way, if they *wanted* to have to values, they could.