Organized steps but identified a flaw
The current chaining means that if an Omni step is added, it impacts requests before
it is added.  This is completely unintuitive.  I suspect I will need to refactor
the Step type to include not just a closure, but an indication of whether it persists.
In this way, the reduction of steps will leave residual steps (omni steps).  I can also
preserve order in that case.

But, as with everything, I need to try it and see if I uncover any flaws.