Got WebSocket subscriptions working
Note that this imports an implementation of WebSocket.  In the browser, this should already
be there (along with EventSource).  So these wouldn't need to be imported in that case.
So the best thing to do here would be to somehow mock these dependencies during testing
so they are imported as if this was a browser.  That way, a bundle generated from the
library wouldn't bundle these implementations.